CARL -> Initialize ResourceManager before loading CA cert, to avoid crash while running database migration.
